Module sel4_cspace::interface

暴露给外部的接口

Re-exports

pub use super::arch::cap_t;

pub use super::arch::CapTag;

Structs

cte_t

由cap_t和 mdb_node 组成，是CSpace的基本组成单元

finaliseCap_ret

mdb_node_t

Functions

cte_insert

将一个cap插入slot中并维护能力派生树

cte_move

将一个cap插入slot中并删除原节点

cte_swap

交换两个slot，并将新的cap数据填入

insert_new_cap

insert a new cap to slot, set parent’s next is slot.

resolve_address_bits

从cspace寻址特定的slot

same_object_as

Check whether two caps point to the same kernel object, if not, whether two kernel objects use the same memory region.